// src/models/DescCen1.ts
export interface DescCen1 {
  IdDescCen1: number;
  nameDescCen1: string | null;
  codeDescCen1: string | null;
}

export interface DescCen1Create {
  nameDescCen1: string | null;
  codeDescCen1: string | null;
}

export interface DescCen1Update {
  nameDescCen1: string | null;
  codeDescCen1: string | null;
}